move media helpers from api to services
