+++ src/service/Cargo.toml
@@ -69,6 +69,7 @@ termimad.optional = true
 tokio.workspace = true
 tracing.workspace = true
 url.workspace = true
+webpage.workspace = true
 
 [lints]
 workspace = true
